Flask-ified OOCK. Charts now load asynchronously
