+++ tools/oock/filter_utils.py
@@ -26,13 +26,13 @@ def select_filter(samples, keys, label_keys, f):
   Args:
     samples: the list of samples
     keys: the sample fields to filter by
-    label_keys: the fields in sample['labels'] to filter by
+    label_keys: the fields in sample['metadata'] to filter by
     f: The lambda. The order of parameters must be as follows: keys, then
        label_keys
   """
   def select_f(s):
     values = [s[key] for key in keys]
-    label_values = [get_label(s, label) for label in label_keys]
+    label_values = [s['metadata'][label] for label in label_keys]
     return f(*(values + label_values))
   samples[:] = [s for s in samples if select_f(s)]
 
@@ -46,13 +46,13 @@ def filter_by_value(samples, key, value):
 def filter_by_label_value(samples, label, value):
   """
   Keeps only samples where the value corresponding to the specified label in 
-  sample['labels'] is equal to the specified value.
+  sample['metadata'] is equal to the specified value.
   """
-  filter_by(samples, 'labels', lambda s: ('|%s:%s|' % (label, value)) in s)
+  filter_by(samples, 'metadata', lambda s: s[label] == value)
 
 def filter_by_label(samples, label, f):
   """
-  Filter samples based on a label in sample['labels'] using a lambda. For each
+  Filter samples based on a label in sample['metadata'] using a lambda. For each
   sample, the label's value will be passed into the lambda. If the lambda
   returns True, the sample will be kept. The list of samples is modified
   in-place.
@@ -69,15 +69,6 @@ def filter_by_label(samples, label, f):
     else:
       return f(match.group(1))
   filter_by(samples, 'labels', label_search)
-
-def get_label(sample, label):
-  """Get the value associated with a label in sample['lables']
-  """
-  match = re.search('\\|%s:([^\\|]*)\\|' % label, sample['labels'])
-  if not match:
-    return None
-  else:
-    return match.group(1)
 
 def one(samples):
   """
